topics/ai_infrastructure/nl_sql_rag/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
outlines>=0.0.34
snowflake-connector-python>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
gradio>=3.40.0
